+++ src/Mangata.ts
@@ -2,7 +2,7 @@ import { ApiPromise } from "@polkadot/api";
 import { KeyringPair } from "@polkadot/keyring/types";
 import { WsProvider } from "@polkadot/rpc-provider/ws";
 import { options } from "@mangata-finance/types";
-import { BN } from "@polkadot/util";
+import { BN, hexToString } from "@polkadot/util";
 
 import { Rpc } from "./services/Rpc";
 import { Tx } from "./services/Tx";
@@ -150,6 +150,90 @@ export class Mangata {
   public async disconnect(): Promise<void> {
     const api = await this.getApi();
     await api.disconnect();
+  }
+
+  public async sendTokenFromParachainToMangata(
+    url: string,
+    tokenSymbol: string,
+    destWeight: string,
+    account: string | KeyringPair,
+    mangataAddress: string,
+    amount: BN,
+    txOptions?: XcmTxOptions
+  ) {
+    const api = await this.getApi();
+    return await Tx.sendTokenFromParachainToMangata(
+      api,
+      url,
+      tokenSymbol,
+      destWeight,
+      account,
+      mangataAddress,
+      amount,
+      txOptions
+    );
+  }
+
+  public async sendTokenFromMangataToParachain(
+    tokenSymbol: string,
+    withWeight: string,
+    parachainId: number,
+    account: string | KeyringPair,
+    destinationAddress: string,
+    amount: BN,
+    txOptions?: XcmTxOptions
+  ) {
+    const api = await this.getApi();
+    return await Tx.sendTokenFromMangataToParachain(
+      api,
+      tokenSymbol,
+      withWeight,
+      parachainId,
+      account,
+      destinationAddress,
+      amount,
+      txOptions
+    );
+  }
+
+  public async sendTokenFromParachainToMangataFee(
+    url: string,
+    tokenSymbol: string,
+    destWeight: string,
+    account: string | KeyringPair,
+    mangataAddress: string,
+    amount: BN
+  ) {
+    const api = await this.getApi();
+    return await Fee.sendTokenFromParachainToMangataFee(
+      api,
+      url,
+      tokenSymbol,
+      destWeight,
+      account,
+      mangataAddress,
+      amount
+    );
+  }
+
+  public async sendTokenFromMangataToParachainFee(
+    tokenSymbol: string,
+    withWeight: string,
+    parachainId: number,
+    account: string | KeyringPair,
+    destinationAddress: string,
+    amount: BN
+  ) {
+    const api = await this.getApi();
+    return await Fee.sendTokenFromMangataToParachainFee(
+      api,
+      tokenSymbol,
+      withWeight,
+      parachainId,
+      account,
+      destinationAddress,
+      amount
+    );
   }
 
   public async sendKusamaTokenFromRelayToParachain(