+++ src/services/Query.ts
@@ -134,10 +134,11 @@ export class Query {
     // we need to filter out ETH and Dummy liquidity token
     // then we need to display symbol for liquidity token
     return Object.values(completeAssetsInfo)
-      .filter((assetsInfo) => assetsInfo.id !== "1" && assetsInfo.id !== "3")
+      .filter((assetsInfo) => !["1", "3", "6"].includes(assetsInfo.id))
       .reduce((obj, item) => {
         const asset = {
           ...item,
+          name: item.name.replace(/0x\w+/, "").replace(/[A-Z]/g, " $&").trim(),
           symbol: item.symbol.includes("TKN")
             ? item.symbol
                 .split("-")