+++ src/utils/getAssetsInfoWithIds.ts
@@ -13,6 +13,7 @@ export const getAssetsInfoWithIds = async (api: ApiPromise) => {
     .reduce((obj, item) => {
       const asset = {
         ...item,
+        name: item.name.replace(/0x\w+/, "").replace(/[A-Z]/g, " $&").trim(),
         symbol: item.symbol.includes("TKN")
           ? item.symbol
               .split("-")